feat: bottom sheet 선택 될 경우 navibar 제거 및 컴포넌트 폴더명 변경 #138
